[Move `@kbn/config-schema` to server] `examples/response_stream`
