+++ examples/response_stream/common/api/reducer_stream/index.ts
@@ -7,5 +7,4 @@
  */
 
 export { reducerStreamReducer } from './reducer';
-export { reducerStreamRequestBodySchema } from './request_body_schema';
-export type { ReducerStreamRequestBodySchema } from './request_body_schema';
+export type { ReducerStreamRequestBodySchema } from '../../../server/routes/schemas/reducer_stream';
