+++ examples/response_stream/server/routes/schemas/reducer_stream/request_body_schema.ts
