+++ examples/response_stream/server/routes/schemas/simple_string_stream/request_body_schema.ts
