+++ examples/response_stream/server/routes/single_string_stream.ts
@@ -9,7 +9,7 @@
 import type { IRouter, Logger } from '@kbn/core/server';
 import { streamFactory } from '@kbn/ml-response-stream/server';
 
-import { simpleStringStreamRequestBodySchema } from '../../common/api/simple_string_stream';
+import { simpleStringStreamRequestBodySchema } from './schemas/simple_string_stream';
 import { RESPONSE_STREAM_API_ENDPOINT } from '../../common/api';
 
 function timeout(ms: number) {
